validacion formularios plan de estudio y obtener fotografia

+++ src/main/java/com/gl05/bad/servicio/PlanEstudioService.java
@@ -1,6 +1,7 @@
 package com.gl05.bad.servicio;
 
 import com.gl05.bad.domain.PlanEstudio;
+import java.math.BigDecimal;
 import org.springframework.data.jpa.datatables.mapping.DataTablesInput;
 import org.springframework.data.jpa.datatables.mapping.DataTablesOutput;
 
@@ -8,7 +9,7 @@ public interface PlanEstudioService {
     
     public DataTablesOutput<PlanEstudio> listarPlanEstudio(DataTablesInput input);
     
-    public void proAgregar(Long idMaestria, String codPlan, String modalidad, long cumMinimo, long notaMinimaAprobacion, short totalAsignaturas, short totalUv, short duracion_carrera, String tituloOtorgar, short anio);
+    public void proAgregar(Long idMaestria, String codPlan, String modalidad, BigDecimal cumMinimo, BigDecimal notaMinimaAprobacion, short totalAsignaturas, short totalUv, short duracion_carrera, String tituloOtorgar, short anio);
     
     public void eliminar(PlanEstudio planEstudio);
